Code cleanup for img_show and image_downloader procedures

+++ src/image_manipulation/image_downloader.py
@@ -4,57 +4,64 @@ Created on 29 ago 2017
 @author: davide
 '''
 
+'''
+    Demonstrates how to download images from a server, how to get them from disk,
+    how to get them shown
+'''
+
 import os
-
 import urllib.request
-
 import ssl
 
-ssl._create_default_https_context = ssl._create_unverified_context
+OVERWRITE = 0
 
+print("Downloading from AWS...")
+ssl._create_default_https_context = ssl._create_unverified_context
 landing_foldername = 'img_align_celeba'
 if not os.path.isdir(landing_foldername): 
     os.mkdir(landing_foldername)
 for img_index in range(1, 11):
     f = '000%03d.jpg' % img_index
     url = 'https://s3.amazonaws.com/cadl/celeb-align/' + f
-    print("url: '" + url + "'")
+    print("\turl: '" + url + "'")
     
     image_filename = os.path.join(landing_foldername, f)
     if os.path.isfile(image_filename):
-        os.remove(image_filename)
+        if OVERWRITE > 0:
+            os.remove(image_filename)
+        else:
+            continue
     urllib.request.urlretrieve(url, image_filename)
     
-print('Download complete.')
+print('\tDownload complete.')
 
+print("Listing images from the disk...")
 files = os.listdir(landing_foldername)
-print(files)
+print("\t" + str(files))
 
 files = [file_i for file_i in os.listdir(landing_foldername) if '.jpg' in file_i]
-print(files)
+print("\t" + str(files))
 
 files = [file_i for file_i in os.listdir(landing_foldername)
         if '.jpg' in file_i or '.png' in file_i or '.jpeg' in file_i]
-print(files)
+print("\t" + str(files))
 
-import matplotlib.pyplot as plt
-import numpy as np
 
+print("Read images from disk...")
 image_filename = os.path.join(landing_foldername, files[0])
-
 import matplotlib.image as mpimg
-plt.imshow(mpimg.imread(image_filename))
+import matplotlib.pyplot as plt
 
+print("Show image...")
 img = plt.imread(image_filename)
 plt.imshow(img)
+plt.show()
 
-print("img.shape: " + str(img.shape) + ".")
+print("\timg.shape: " + str(img.shape) + ".")
 
+print("Show RED channel only...")
 plt.figure()
-plt.imgshow(img[:, :, 0])
-plt.figure()
-
 plt.imshow(img[:, :, 0])
-plt.figure()
+plt.show()
 
 print('Done.')
